refactor(test): de-dupe test boilerplate + clear scanChunk complexity (fallow:audit)
Drive the fallow:audit lane down via genuine de-dup (no accepting baseline exists
— the gate is a zero-introduced ratchet). Complexity introduced 1 -> 0 and
duplication introduced 15 -> 9:

- scanChunk: suppress the cyclomatic/cognitive finding with fallow's documented
  inline directive — it is a byte-for-byte POC port locked down by a fast-check
  property test; splitting it would scatter the state machine without reducing
  real complexity.
- chat tests: import the shared `provideScopedLayer` from @beep/test-utils
  (delete local copies) — dissolves the large cross-repo test-structure clone.
- pglite tests: extract `makePgliteIntegrationGate` into @beep/test-utils (env
  gate + layer, no @beep/postgres dep) and adopt it in the new pglite tests.
- tables tests: adopt the existing @beep/test-utils `systemPrincipal` /
  `baseEntityFixtureInput` fixtures instead of local copies.

The residual 9 duplication groups are structurally irreducible from this branch
(per-package vitest.config.ts clones with no fallow glob-ignore; cross-package
test-harness clones needing 5+ existing packages refactored) and the yeet audit
invocation passes no baseline flag — a repo-tooling/policy matter owned by the
fallow-quality-enforcement packet. See
goals/desktop-chat-surface/history/2026-06-14-fallow-audit-investigation.md.

All touched packages green: check + test + lint, integration lane 3/3.
