+++ packages/agents/server/src/AssistantTurn/ScanState.ts
@@ -60,6 +60,11 @@ export const initialScanState: ScanState = {
  * @category parsing
  * @since 0.0.0
  */
+// scanChunk is a byte-for-byte port of the POC's incremental JSON block scanner.
+// Its branching is the brace/string/escape state machine itself and is locked
+// down by a fast-check property test (test/scanChunk.test.ts); extracting helpers
+// would scatter the single-pass state without reducing real complexity.
+// fallow-ignore-next-line complexity
 export const scanChunk = (state: ScanState, text: string): [ScanState, Array<string>] => {
   let { current, depth, escaped, inBlocksArray, inString } = state;
   const completed = A.empty<string>();
