+++ packages/foundation/modeling/md/.beep/repo-exports/catalog.shard.jsonc
@@ -13,7 +13,7 @@
   },
   "fingerprint": {
     "algorithm": "sha256",
-    "digest": "4356c592ff600a6cd1c2ae3c483cf6d0a6c30e1235b84c236aae0396c20465e0",
+    "digest": "1577652f20a930a95a332d5d828f0654f650006013c9c992e828ef9e55df4c93",
     "inputs": [
       {
         "path": "generator:repo-exports-catalog",
@@ -32,8 +32,8 @@
       },
       {
         "path": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sha256": "7778f4bf8af7d1be64d5e3ba4fc9574b2e30d93f0cb1287f6ad382d28ac6c234",
-        "bytes": 31606
+        "sha256": "08e15214b8bd0be50c09d0b6365a275faa491f7c09e6b6cd2c7e9e967b2edd5e",
+        "bytes": 32246
       },
       {
         "path": "packages/foundation/modeling/md/src/Md.render.ts",
@@ -169,7 +169,7 @@
         "symbolName": "Block",
         "exportKind": "const",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1503,
+        "sourceLine": 1516,
         "summary": "Discriminated union of block Markdown AST nodes.",
         "categories": [
           "models"
@@ -194,7 +194,7 @@
         "symbolName": "Block",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1524,
+        "sourceLine": 1537,
         "summary": "Companion namespace for {@link Block}.",
         "categories": [
           "models"
@@ -218,7 +218,7 @@
         "symbolName": "Block",
         "exportKind": "type",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1516,
+        "sourceLine": 1529,
         "summary": "Runtime type for {@link Block}.",
         "categories": [
           "models"
@@ -686,7 +686,7 @@
         "symbolName": "Document",
         "exportKind": "class",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1578,
+        "sourceLine": 1591,
         "summary": "Root Markdown document AST.",
         "categories": [
           "models"
@@ -711,7 +711,7 @@
         "symbolName": "Document",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1596,
+        "sourceLine": 1609,
         "summary": "Companion namespace for {@link Document}.",
         "categories": [
           "models"
@@ -1478,7 +1478,7 @@
         "symbolName": "Hr",
         "exportKind": "class",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1459,
+        "sourceLine": 1472,
         "summary": "Horizontal rule block.",
         "categories": [
           "models"
@@ -1503,7 +1503,7 @@
         "symbolName": "Hr",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1473,
+        "sourceLine": 1486,
         "summary": "Companion namespace for {@link Hr}.",
         "categories": [
           "models"
@@ -2317,7 +2317,7 @@
         "symbolName": "Pre",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1429,
+        "sourceLine": 1434,
         "summary": "Companion namespace for {@link Pre}.",
         "categories": [
           "models"
@@ -4431,7 +4431,7 @@
         "symbolName": "Block",
         "exportKind": "const",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1503,
+        "sourceLine": 1516,
         "summary": "Discriminated union of block Markdown AST nodes.",
         "categories": [
           "models"
@@ -4456,7 +4456,7 @@
         "symbolName": "Block",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1524,
+        "sourceLine": 1537,
         "summary": "Companion namespace for {@link Block}.",
         "categories": [
           "models"
@@ -4480,7 +4480,7 @@
         "symbolName": "Block",
         "exportKind": "type",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1516,
+        "sourceLine": 1529,
         "summary": "Runtime type for {@link Block}.",
         "categories": [
           "models"
@@ -4748,7 +4748,7 @@
         "symbolName": "Document",
         "exportKind": "class",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1578,
+        "sourceLine": 1591,
         "summary": "Root Markdown document AST.",
         "categories": [
           "models"
@@ -4773,7 +4773,7 @@
         "symbolName": "Document",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1596,
+        "sourceLine": 1609,
         "summary": "Companion namespace for {@link Document}.",
         "categories": [
           "models"
@@ -5140,7 +5140,7 @@
         "symbolName": "Hr",
         "exportKind": "class",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1459,
+        "sourceLine": 1472,
         "summary": "Horizontal rule block.",
         "categories": [
           "models"
@@ -5165,7 +5165,7 @@
         "symbolName": "Hr",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1473,
+        "sourceLine": 1486,
         "summary": "Companion namespace for {@link Hr}.",
         "categories": [
           "models"
@@ -5579,7 +5579,7 @@
         "symbolName": "Pre",
         "exportKind": "namespace",
         "sourcePath": "packages/foundation/modeling/md/src/Md.model.ts",
-        "sourceLine": 1429,
+        "sourceLine": 1434,
         "summary": "Companion namespace for {@link Pre}.",
         "categories": [
           "models"
