+++ packages/tooling/test-kit/test-utils/.beep/repo-exports/catalog.shard.jsonc
@@ -13,7 +13,7 @@
   },
   "fingerprint": {
     "algorithm": "sha256",
-    "digest": "2fc8a32b6709fa4c631f2de9db5fa99825c0585001c6ca9d616de47f27255af6",
+    "digest": "018a8c2ffb76cd17a083309b750161d1c50fb092876190dccd6bf1e7081a611b",
     "inputs": [
       {
         "path": "generator:repo-exports-catalog",
@@ -52,8 +52,8 @@
       },
       {
         "path": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
-        "sha256": "a438ad6be256746840ef5e09039ed2ea12a97a5cb07e2487b7b89ee7107bbdda",
-        "bytes": 39754
+        "sha256": "206b7ac8907a32eb663e536de665079af3aa5f532c56c1cb895103f96e424f58",
+        "bytes": 42205
       },
       {
         "path": "packages/tooling/tool/cli/src/commands/Quality/internal/QualityArtifactSupport.ts",
@@ -81,8 +81,8 @@
       "@beep/test-utils/SqlTest"
     ],
     "counts": {
-      "publicExportEntries": 69,
-      "uniqueSymbols": 23,
+      "publicExportEntries": 72,
+      "uniqueSymbols": 24,
       "sourceFiles": 5
     },
     "exports": [
@@ -170,11 +170,11 @@
         "importSpecifier": "@beep/test-utils",
         "exportSubpath": ".",
         "exportedFromPath": "packages/tooling/test-kit/test-utils/src/index.ts",
-        "symbolName": "makePgliteSqlTestLayer",
+        "symbolName": "makePgliteIntegrationGate",
         "exportKind": "const",
         "sourcePath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
-        "sourceLine": 1131,
-        "summary": "Build the recommended PGLite SQL test layer for vertical-slice integration tests.",
+        "sourceLine": 1167,
+        "summary": "Pre-computed gate values for PGLite integration tests.",
         "categories": [
           "constructors"
         ],
@@ -182,13 +182,28 @@
           "0.0.0"
         ],
         "tags": [
-          "@param",
-          "@returns",
           "@example",
           "@category",
           "@since"
         ],
-        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils . makepglitesqltestlayer const build the recommended pglite sql test layer for vertical-slice integration tests. constructors packages/tooling/test-kit/test-utils/src/sqltest.ts"
+        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils . makepgliteintegrationgate const pre-computed gate values for pglite integration tests. constructors packages/tooling/test-kit/test-utils/src/sqltest.ts"
+      },
+      {
+        "packageName": "@beep/test-utils",
+        "packagePath": "packages/tooling/test-kit/test-utils",
+        "topoOrder": 0,
+        "importSpecifier": "@beep/test-utils",
+        "exportSubpath": ".",
+        "exportedFromPath": "packages/tooling/test-kit/test-utils/src/index.ts",
+        "symbolName": "makePgliteSqlTestLayer",
+        "exportKind": "const",
+        "sourcePath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
+        "sourceLine": 1197,
+        "summary": "",
+        "categories": [],
+        "since": [],
+        "tags": [],
+        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils . makepglitesqltestlayer const packages/tooling/test-kit/test-utils/src/sqltest.ts"
       },
       {
         "packageName": "@beep/test-utils",
@@ -806,11 +821,11 @@
         "importSpecifier": "@beep/test-utils/index",
         "exportSubpath": "./index",
         "exportedFromPath": "packages/tooling/test-kit/test-utils/src/index.ts",
-        "symbolName": "makePgliteSqlTestLayer",
+        "symbolName": "makePgliteIntegrationGate",
         "exportKind": "const",
         "sourcePath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
-        "sourceLine": 1131,
-        "summary": "Build the recommended PGLite SQL test layer for vertical-slice integration tests.",
+        "sourceLine": 1167,
+        "summary": "Pre-computed gate values for PGLite integration tests.",
         "categories": [
           "constructors"
         ],
@@ -818,13 +833,28 @@
           "0.0.0"
         ],
         "tags": [
-          "@param",
-          "@returns",
           "@example",
           "@category",
           "@since"
         ],
-        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils/index ./index makepglitesqltestlayer const build the recommended pglite sql test layer for vertical-slice integration tests. constructors packages/tooling/test-kit/test-utils/src/sqltest.ts"
+        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils/index ./index makepgliteintegrationgate const pre-computed gate values for pglite integration tests. constructors packages/tooling/test-kit/test-utils/src/sqltest.ts"
+      },
+      {
+        "packageName": "@beep/test-utils",
+        "packagePath": "packages/tooling/test-kit/test-utils",
+        "topoOrder": 0,
+        "importSpecifier": "@beep/test-utils/index",
+        "exportSubpath": "./index",
+        "exportedFromPath": "packages/tooling/test-kit/test-utils/src/index.ts",
+        "symbolName": "makePgliteSqlTestLayer",
+        "exportKind": "const",
+        "sourcePath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
+        "sourceLine": 1197,
+        "summary": "",
+        "categories": [],
+        "since": [],
+        "tags": [],
+        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils/index ./index makepglitesqltestlayer const packages/tooling/test-kit/test-utils/src/sqltest.ts"
       },
       {
         "packageName": "@beep/test-utils",
@@ -1392,11 +1422,11 @@
         "importSpecifier": "@beep/test-utils/SqlTest",
         "exportSubpath": "./SqlTest",
         "exportedFromPath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
-        "symbolName": "makePgliteSqlTestLayer",
+        "symbolName": "makePgliteIntegrationGate",
         "exportKind": "const",
         "sourcePath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
-        "sourceLine": 1131,
-        "summary": "Build the recommended PGLite SQL test layer for vertical-slice integration tests.",
+        "sourceLine": 1167,
+        "summary": "Pre-computed gate values for PGLite integration tests.",
         "categories": [
           "constructors"
         ],
@@ -1404,13 +1434,28 @@
           "0.0.0"
         ],
         "tags": [
-          "@param",
-          "@returns",
           "@example",
           "@category",
           "@since"
         ],
-        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils/sqltest ./sqltest makepglitesqltestlayer const build the recommended pglite sql test layer for vertical-slice integration tests. constructors packages/tooling/test-kit/test-utils/src/sqltest.ts"
+        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils/sqltest ./sqltest makepgliteintegrationgate const pre-computed gate values for pglite integration tests. constructors packages/tooling/test-kit/test-utils/src/sqltest.ts"
+      },
+      {
+        "packageName": "@beep/test-utils",
+        "packagePath": "packages/tooling/test-kit/test-utils",
+        "topoOrder": 0,
+        "importSpecifier": "@beep/test-utils/SqlTest",
+        "exportSubpath": "./SqlTest",
+        "exportedFromPath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
+        "symbolName": "makePgliteSqlTestLayer",
+        "exportKind": "const",
+        "sourcePath": "packages/tooling/test-kit/test-utils/src/SqlTest.ts",
+        "sourceLine": 1197,
+        "summary": "",
+        "categories": [],
+        "since": [],
+        "tags": [],
+        "searchText": "@beep/test-utils packages/tooling/test-kit/test-utils @beep/test-utils/sqltest ./sqltest makepglitesqltestlayer const packages/tooling/test-kit/test-utils/src/sqltest.ts"
       },
       {
         "packageName": "@beep/test-utils",
